+++ mcdc/mcdc_get/tally.py
@@ -185,59 +185,59 @@ def bin_chunk(start, length, tally, data):
 
 
 @njit
-def bin_sum(index, tally, data):
-    offset = tally["bin_sum_offset"]
+def bin_mean(index, tally, data):
+    offset = tally["bin_mean_offset"]
     return data[offset + index]
 
 
 @array_return(nb.types.float64)
-def bin_sum_all(tally, data):
-    start = tally["bin_sum_offset"]
-    size = tally["bin_sum_length"]
+def bin_mean_all(tally, data):
+    start = tally["bin_mean_offset"]
+    size = tally["bin_mean_length"]
     end = start + size
     return array_result(data[start:end])
 
 
 @njit
-def bin_sum_last(tally, data):
-    start = tally["bin_sum_offset"]
-    size = tally["bin_sum_length"]
+def bin_mean_last(tally, data):
+    start = tally["bin_mean_offset"]
+    size = tally["bin_mean_length"]
     end = start + size
     return data[end - 1]
 
 
 @array_return(nb.types.float64)
-def bin_sum_chunk(start, length, tally, data):
-    start += tally["bin_sum_offset"]
+def bin_mean_chunk(start, length, tally, data):
+    start += tally["bin_mean_offset"]
     end = start + length
     return array_result(data[start:end])
 
 
 @njit
-def bin_sum_square(index, tally, data):
-    offset = tally["bin_sum_square_offset"]
+def bin_sum_squared_deviations(index, tally, data):
+    offset = tally["bin_sum_squared_deviations_offset"]
     return data[offset + index]
 
 
 @array_return(nb.types.float64)
-def bin_sum_square_all(tally, data):
-    start = tally["bin_sum_square_offset"]
-    size = tally["bin_sum_square_length"]
+def bin_sum_squared_deviations_all(tally, data):
+    start = tally["bin_sum_squared_deviations_offset"]
+    size = tally["bin_sum_squared_deviations_length"]
     end = start + size
     return array_result(data[start:end])
 
 
 @njit
-def bin_sum_square_last(tally, data):
-    start = tally["bin_sum_square_offset"]
-    size = tally["bin_sum_square_length"]
+def bin_sum_squared_deviations_last(tally, data):
+    start = tally["bin_sum_squared_deviations_offset"]
+    size = tally["bin_sum_squared_deviations_length"]
     end = start + size
     return data[end - 1]
 
 
 @array_return(nb.types.float64)
-def bin_sum_square_chunk(start, length, tally, data):
-    start += tally["bin_sum_square_offset"]
+def bin_sum_squared_deviations_chunk(start, length, tally, data):
+    start += tally["bin_sum_squared_deviations_offset"]
     end = start + length
     return array_result(data[start:end])